updating img component quality
